pyproject.toml
[project]
name = "simplemodels"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ase>=3.27.0",
    "click>=8.3.1",
    "minimahopping",
    "orb-models>=0.6.2",
    "pandas>=3.0.1",
    "pyyaml>=6.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "pytest>=9.0.2",
    "pytest-coverage>=0.0",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.uv.sources]
minimahopping = { git = "https://gitlab.com/goedeckergroup/ase_mh.git" }

[project.scripts]
example_app = "simplemodels.basic:main"
